Separate specialized and generic kernels
