+++ src/amd64/kernel_amd64_avx512f.cpp
@@ -6,5 +6,5 @@
 #endif
 
 #define QUANT_KERNEL_IMPL install_quant_amd64_avx512f
-#include "../kernels.inl"
+#include "../kernels/kernels.inl"
 #undef QUANT_KERNEL_IMPL
