+++ src/amd64/kernel_amd64_sse42.cpp
@@ -8,5 +8,5 @@
 #endif
 
 #define QUANT_KERNEL_IMPL install_quant_amd64_sse42
-#include "../kernels.inl"
+#include "../kernels/kernels.inl"
 #undef QUANT_KERNEL_IMPL
