+++ src/kernel_generic.cpp
@@ -1,3 +1,3 @@
 #define QUANT_KERNEL_IMPL install_quant_generic
-#include "kernels.inl"
+#include "kernels/kernels_generic.inl"
 #undef QUANT_KERNEL_IMPL
